server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
PyMySQL==1.1.1
python-dotenv==1.1.0
PyJWT==2.10.1
bcrypt==4.3.0
